+++ packages/signalwire/src/VideoChat/components/Chat/types.ts
@@ -1,0 +1,7 @@
+import { FlexProps } from '@cmpsr/components';
+
+export interface ChatProps extends FlexProps {
+  userId: string;
+  floating?: boolean;
+  channel: string;
+}