+++ packages/signalwire/src/VideoChat/components/VideoChat/index.ts
@@ -1,0 +1,2 @@
+export * from './VideoChat';
+export * from './types';