+++ packages/signalwire/src/VideoChat/utils/client/isRoomAvailable/isRoomAvailable.ts
@@ -1,0 +1,6 @@
+import { apiClient } from '../../api';
+
+export const isRoomAvailable = async (room: string) => {
+  const response = await apiClient.get(`/video/rooms/${room}`);
+  return response.data?.room === null;
+};